requirements.txt
Flask==2.3.3
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
Pillow==10.0.1
opencv-python==4.8.1.78
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
Werkzeug==2.3.7
gunicorn==21.2.0
